Add files via upload
Implemented a console-based Email Management System in C++ as part of a software project. This system provides users with functionalities such as user signup, sign-in, inbox, outbox, composing and sending messages, and signing out. Leveraging file handling, the project stores user credentials and email messages, allowing users to interact with their emails in a simple and intuitive way.